EPMDJ-3557: Update Overview page of non-latest build
Issue: EPMDJ-3557
