--- src/pages/agent-full-page/coverage-plugin/code-coverage-card/detailed-code-coverage-card/coverage-distribution/coverage-distribution.module.scss
@@ -1,83 +1,0 @@
-@import '~common/style-guide/index';
-
-.coverage-distribution {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  flex-grow: 1;
-  color: $slate-gray;
-
-  &__distribution-container {
-    display: flex;
-    align-items: center;
-    flex-grow: 1;
-
-    & > :not(:last-child) {
-      margin-right: 40px;
-    }
-  }
-
-  &__coverage-item {
-    display: flex;
-    flex-direction: column;
-    flex-grow: 1;
-
-    & > :not(:last-child) {
-      border-bottom: 1px solid map-get($color-palette, light-blue);
-    }
-  }
-
-  &__item-name {
-    margin-left: 8px;
-    font-size: 12px;
-    line-height: 32px;
-    text-transform: capitalize;
-  }
-
-  &__recording-wrapper {
-    margin-left: 8px;
-  }
-
-  &__recording-icon {
-    height: 8px;
-    width: 8px;
-    border-radius: 4px;
-    background-color: map-get($color-palette, red100);
-    margin-right: 3px;
-    margin-left: 4px;
-    animation-name: blinker;
-    animation-iteration-count: infinite;
-    animation-timing-function: cubic-bezier(1, 0, 0, 1);
-    animation-duration: 1s;
-
-    @keyframes blinker {
-      from {
-        opacity: 1;
-      }
-      to {
-        opacity: 0;
-      }
-    }
-  }
-
-  &__recording-text {
-    font-size: 12px;
-    line-height: 12px;
-    color: map-get($color-palette, red100);
-    font-family: $semibold-font;
-  }
-
-  &__tests-count {
-    @extend %tests-styles;
-  }
-
-  &__test-type-coverage {
-    @extend %tests-styles;
-  }
-}
-
-%tests-styles {
-  font-family: $semibold-font;
-  font-size: 12px;
-  color: map-get($color-palette, black100);
-}
