--- src/pages/agent-full-page/coverage-plugin/code-coverage-card/detailed-code-coverage-card/coverage-distribution/index.ts
@@ -1,1 +1,0 @@
-export { CoverageDistribution } from './coverage-distribution';
