--- src/pages/agent-full-page/coverage-plugin/code-coverage-card/index.ts
@@ -1,1 +1,0 @@
-export { DetailedCodeCoverageCard } from './detailed-code-coverage-card';
