+++ src/pages/agent-full-page/coverage-plugin/overview/active-build-coverage-info/active-build-coverage-info.module.scss
@@ -1,0 +1,63 @@
+@import '~common/style-guide/index';
+
+.active-build-coverage-info {
+  width: calc(100% - 258px - 52px);
+  height: 100%;
+  font-size: 12px;
+  line-height: 16px;
+  color: map-get($monochrome, default);
+
+  &__title {
+    font-family: $semibold-font;
+  }
+
+  &__build-coverage-status {
+    display: flex;
+    align-items: baseline;
+    margin: 24px 0 12px 0;
+    font-size: 12px;
+  }
+
+  &__build-coverage-type {
+    &::before {
+      content: '';
+      display: inline-block;
+      margin-right: 4px;
+      width: 8px;
+      height: 8px;
+      border-radius: 4px;
+    }
+
+    &:nth-child(1) {
+      &::before {
+        background-color: map-get($data-visualization, build-cover);
+      }
+    }
+
+    &:nth-child(2) {
+      margin-left: 16px;
+      &::before {
+        background-color: map-get($data-visualization, overlapping);
+      }
+    }
+
+    &:nth-child(3) {
+      margin-left: 16px;
+      &::before {
+        background-color: map-get($data-visualization, scope-cover);
+      }
+    }
+  }
+
+  &__build-coverage-percentage {
+    margin-right: 8px;
+    font-size: 32px;
+    line-height: 40px;
+    color: map-get($monochrome, black);
+  }
+
+  &__unique-coverage-percentage {
+    line-height: 24px;
+    color: map-get($data-visualization, build-cover);
+  }
+}
