+++ src/pages/agent-full-page/coverage-plugin/overview/active-build-coverage-info/index.tsx
@@ -1,0 +1,1 @@
+export { ActiveBuildCoverageInfo } from './active-build-coverage-info';
