+++ src/pages/agent-full-page/coverage-plugin/overview/active-build-coverage-info/multi-progress-bar/multi-progress-bar.tsx
