+++ src/pages/agent-full-page/coverage-plugin/overview/build-coverage-info/build-coverage-info.tsx
@@ -1,96 +1,57 @@
 import * as React from 'react';
+import { NavLink, useParams } from 'react-router-dom';
 import { BEM } from '@redneckz/react-bem-helper';
-import { Panel, Legends } from '@drill4j/ui-kit';
+import { Panel, Legends, MainProgressBar } from '@drill4j/ui-kit';
 
 import { percentFormatter } from 'utils';
-import { ActiveScope } from 'types/active-scope';
 import { BuildCoverage } from 'types/build-coverage';
-import { MultiProgressBar } from './multi-progress-bar';
 
 import styles from './build-coverage-info.module.scss';
 
 interface Props {
   className?: string;
   buildCoverage: BuildCoverage;
-  scope: ActiveScope | null;
-  status?: 'ONLINE' | 'NOT_REGISTERED' | 'BUSY' | 'OFFLINE';
-  loading: boolean;
 }
 
 const buildCoverageInfo = BEM(styles);
 
-export const BuildCoverageInfo = buildCoverageInfo(({
-  className, buildCoverage, scope, status = 'BUSY', loading,
-}: Props) => {
-  const {
-    coverage: {
-      percentage: coveragePercentage = 0,
-      overlap: { percentage: overlapPercentage = 0, methodCount: { covered: overlapCoveredMethods = 0 } = {} } = {},
-    } = {},
-  } = scope || {};
-  const {
-    prevBuildVersion = '',
-    diff = 0,
-    percentage: buildCodeCoverage = 0,
-    finishedScopesCount = 0,
-    methodCount: { covered: buildCoveredMethods = 0 } = {},
-  } = buildCoverage;
-  const uniqueMethods = buildCoveredMethods - overlapCoveredMethods;
+export const BuildCoverageInfo = buildCoverageInfo(({ className, buildCoverage }: Props) => {
+  const { prevBuildVersion = '', diff = 0, percentage: buildCodeCoverage = 0 } = buildCoverage;
+  const { agentId, buildVersion, pluginId } = useParams<{agentId: string, buildVersion: string, pluginId: string }>();
   return (
     <div className={className}>
       <Panel align="space-between">
         <Title data-test="build-coverage-info:title">BUILD COVERAGE</Title>
-        <span>
-          <BuildCoverageType type="build">Build</BuildCoverageType>
-          <BuildCoverageType type="overlapping">Build / Scope overlap</BuildCoverageType>
-          <BuildCoverageType type="scope">Scope</BuildCoverageType>
-        </span>
+        <Link
+          to={`/full-page/${agentId}/${buildVersion}/${pluginId}/scopes/`}
+          data-test="build-coverage-info:all-scopes-link"
+        >
+          All scopes
+        </Link>
       </Panel>
-      <BuildCoverageStatus data-test="build-coverage-info:status">
-        <BuildCoveragePercentage data-test="overview:build-coverage-percentage">
+      <DetailedCodeCoverageInfo data-test="build-coverage-info:detailed-code-coverage-info">
+        <BuildCoveragePercentage data-test="build-coverage-info:build-coverage-percentage">
           {percentFormatter(buildCodeCoverage)}%
         </BuildCoveragePercentage>
-        {finishedScopesCount > 0 && (
-          <Panel align="space-between">
-            {prevBuildVersion && (
-              <span data-test="build-coverage-info:comparing">
-                <b>
-                  {diff >= 0 ? '+ ' : '- '}
-                  {percentFormatter(Math.abs(diff))}%
-                  &nbsp;
-                </b>
-                сomparing to Build:&nbsp;
-                {prevBuildVersion}
-              </span>
-            )}
-            {Boolean(coveragePercentage) && (
-              <span>
-                <UniqueCoveragePercentage data-test="overview:unique-coverage-percentage">
-                  +{percentFormatter(coveragePercentage - overlapPercentage)}%
-                </UniqueCoveragePercentage>
-                &nbsp;in active scope
-              </span>
-            )}
-          </Panel>
+        {prevBuildVersion && (
+          <span data-test="build-coverage-info:comparing">
+            <b>
+              {diff >= 0 ? '+ ' : '- '}
+              {percentFormatter(Math.abs(diff))}%
+              &nbsp;
+            </b>
+            сomparing to Build:&nbsp;
+            {prevBuildVersion}
+          </span>
         )}
-        {status === 'BUSY' && 'Loading...'}
-        {(finishedScopesCount === 0 && status === 'ONLINE') &&
-            'Press “Complete active scope” button to add your scope coverage to the build.'}
-      </BuildCoverageStatus>
-      <MultiProgressBar
-        buildCodeCoverage={buildCodeCoverage}
-        uniqueCodeCoverage={coveragePercentage - overlapPercentage}
-        overlappingCode={overlapPercentage}
-        methods={{ overlapCoveredMethods, buildCoveredMethods, uniqueMethods }}
-        active={loading}
-      />
+      </DetailedCodeCoverageInfo>
+      <MainProgressBar type="primary" value={`${buildCodeCoverage}%`} />
       <Legends />
     </div>
   );
 });
 
 const Title = buildCoverageInfo.title('div');
-const BuildCoverageStatus = buildCoverageInfo.buildCoverageStatus('div');
-const BuildCoverageType = buildCoverageInfo.buildCoverageType('span');
+const DetailedCodeCoverageInfo = buildCoverageInfo.detailedCodeCoverageInfo('div');
 const BuildCoveragePercentage = buildCoverageInfo.buildCoveragePercentage('div');
-const UniqueCoveragePercentage = buildCoverageInfo.uniqueCoveragePercentage('span');
+const Link = buildCoverageInfo.link(NavLink);
