+++ src/pages/agent-full-page/coverage-plugin/overview/build-coverage-info/index.ts
