+++ src/pages/agent-full-page/coverage-plugin/overview/overview.tsx
@@ -1,12 +1,8 @@
 import * as React from 'react';
 import { BEM } from '@redneckz/react-bem-helper';
-import { Link, useParams } from 'react-router-dom';
-import {
-  Panel, Icons,
-} from '@drill4j/ui-kit';
+import { Panel, Icons } from '@drill4j/ui-kit';
 
 import { TabsPanel, Tab } from 'components';
-import { percentFormatter } from 'utils';
 import { BuildCoverage } from 'types/build-coverage';
 import { ActiveScope } from 'types/active-scope';
 import { Methods } from 'types/methods';
@@ -15,10 +11,10 @@ import { TableActionsProvider } from 'modules';
 import { CoveragePluginHeader } from '../coverage-plugin-header';
 import { useBuildVersion } from '../use-build-version';
 import { CoverageDetails } from '../coverage-details';
-import { DetailedCodeCoverageCard } from '../code-coverage-card';
 import { ProjectMethodsCards } from '../project-methods-cards';
 import { usePluginState } from '../../store';
 import { Tests } from '../tests';
+import { ActiveBuildCoverageInfo } from './active-build-coverage-info';
 import { BuildCoverageInfo } from './build-coverage-info';
 import { ActiveScopeInfo } from './active-scope-info';
 
@@ -32,50 +28,25 @@ const overview = BEM(styles);
 
 export const Overview = overview(({ className }: Props) => {
   const [selectedTab, setSelectedTab] = React.useState('methods');
-  const { agentId, buildVersion, loading } = usePluginState();
+  const { agentId, loading } = usePluginState();
   const { status } = useAgent(agentId) || {};
   const buildCoverage = useBuildVersion<BuildCoverage>('/build/coverage') || {};
-  const {
-    prevBuildVersion = '',
-    diff = 0,
-    finishedScopesCount = 0,
-    packageCount: { total = 0 } = {},
-  } = buildCoverage;
   const scope = useBuildVersion<ActiveScope>('/active-scope');
   const buildMethods = useBuildVersion<Methods>('/build/methods') || {};
-  const { pluginId } = useParams<{ pluginId: string}>();
 
   return (
     <div className={className}>
       <CoveragePluginHeader />
       <SummaryPanel align="space-between">
         {scope?.active ? (
-          <BuildCoverageInfo buildCoverage={buildCoverage} scope={scope} status={status} loading={loading} />
+          <>
+            <ActiveBuildCoverageInfo buildCoverage={buildCoverage} scope={scope} status={status} loading={loading} />
+            <ActiveScopeInfo scope={scope} />
+          </>
         )
           : (
-            <DetailedCodeCoverageCard
-              header={(
-                <Panel align="space-between">
-                  Build Coverage
-                  <ScopesListLink to={`/full-page/${agentId}/${buildVersion}/${pluginId}/scopes`}>
-                    {`All Scopes (${finishedScopesCount}) >`}
-                  </ScopesListLink>
-                </Panel>
-              )}
-              coverage={buildCoverage}
-              additionalInfo={(
-                <Panel>
-                  {prevBuildVersion
-                  && `${diff >= 0 ? '+ ' : '- '} ${percentFormatter(
-                    Math.abs(diff),
-                  )}% comparing to Build: ${prevBuildVersion}`}
-                </Panel>
-              )}
-            />
+            <BuildCoverageInfo buildCoverage={buildCoverage} />
           )}
-        {scope?.active && (
-          <ActiveScopeInfo scope={scope} />
-        )}
       </SummaryPanel>
       <RoutingTabsPanel>
         <TabsPanel activeTab={selectedTab} onSelect={setSelectedTab}>
@@ -102,7 +73,7 @@ export const Overview = overview(({ className }: Props) => {
                 topic="/build/coverage/packages"
                 associatedTestsTopic="/build/associated-tests"
                 classesTopicPrefix="build"
-                packageCount={total}
+                packageCount={buildCoverage.packageCount?.total}
               />
             </TableActionsProvider>
           </>
@@ -115,7 +86,6 @@ export const Overview = overview(({ className }: Props) => {
 });
 
 const SummaryPanel = overview.summaryPanel(Panel);
-const ScopesListLink = overview.scopesListLink(Link);
 const RoutingTabsPanel = overview.routingTabsPanel('div');
 const TabContent = overview.tabContent('div');
 const TabIconWrapper = overview.tabIconWrapper('div');
